editing tdown.fish

+++ tdown.fish
@@ -1,8 +1,12 @@
-# Defined in /home/michael/.config/fish/buffer/tdown.fish @ line 2
+# Defined in /usr/home/michael/.config/fish/buffer/tdown.fish @ line 2
 function tdown
-	set info (prompt 'title @ duration')
-  set title  (echo $info | cut -d @ -f1| trim)
-  set duration (echo $info | cut -d @ -f2 | trim)
-  eval kitty -o font_size=50 -e termdown --no-figlet -v en-us -T (quote $title) $duration
-  echo $title completed at (date) > /tmp/timers-complete.txt
+		if not exists $argv
+				last-timer
+				return 0
+		end
+		set info (prompt 'title @ duration')
+		set title  (echo $info | cut -d @ -f1| trim)
+		set duration (echo $info | cut -d @ -f2 | trim)
+		eval kitty -o font_size=50 -e termdown --no-figlet -v en-us -T (quote $title) $duration
+		echo $title completed at (date) > /tmp/timers-complete.txt
 end
